Revert "Rename library files according to PlatformIO requirements"
This reverts commit 8f8666d0756ccf553a39a90457066baefc781126.
