+++ src/Stepper.cpp
@@ -1,4 +1,4 @@
-#include "STM32Step.h"
+#include "Stepper.h"
 #include "core_pins.h"
 
 Stepper::Stepper(const int _stepPin, const int _dirPin)
